code formatting (py_src)

+++ py_src/diffuserRapper.py
@@ -13,30 +13,32 @@ generate_progress_callback = None
 loadedModelId = None
 loadedVaeModelId = None
 
-def loadPipeline(modelName, torch_dtype, vaeName = None):
+
+def loadPipeline(modelName, torch_dtype, vaeName=None):
     if vaeName is None:
         pipeline = StableDiffusionPipeline.from_pretrained(
-            pretrained_model_name_or_path = modelName,
-            torch_dtype = torch_dtype
+            pretrained_model_name_or_path=modelName,
+            torch_dtype=torch_dtype
         )
     else:
         pipeline = StableDiffusionPipeline.from_pretrained(
-            pretrained_model_name_or_path = modelName,
-            torch_dtype = torch_dtype,
-            vae = AutoencoderKL.from_pretrained(
-                pretrained_model_name_or_path = vaeName, 
-                torch_dtype = torch_dtype
+            pretrained_model_name_or_path=modelName,
+            torch_dtype=torch_dtype,
+            vae=AutoencoderKL.from_pretrained(
+                pretrained_model_name_or_path=vaeName,
+                torch_dtype=torch_dtype
             )
         )
     return pipeline.to("cuda")
 
-def load(modelId, torch_dtype, vaeId = None):
+
+def load(modelId, torch_dtype, vaeId=None):
     global pipe, loadedModelId, loadedVaeModelId
     print('start model load.')
     load_time = time.time()
     pipe = loadPipeline(
-        os.path.join(get_models_path(), modelId), 
-        torch_dtype, 
+        os.path.join(get_models_path(), modelId),
+        torch_dtype,
         None if vaeId == None else os.path.join(get_models_path(), vaeId)
     )
     pipe.safety_checker = lambda images, **kwargs: (images, False)
@@ -46,21 +48,26 @@ def load(modelId, torch_dtype, vaeId = None):
     loadedVaeModelId = vaeId
     print(f"Models loaded in {time_load:.2f}s")
 
+
 def TestLoad(path: str, importType: ModelType):
     try:
         if (importType == ModelType.Model):
-            pipe = StableDiffusionPipeline.from_pretrained(pretrained_model_name_or_path = path)
+            pipe = StableDiffusionPipeline.from_pretrained(
+                pretrained_model_name_or_path=path)
             pipe = None
         elif (importType == ModelType.Vae):
-            vae = AutoencoderKL.from_pretrained(pretrained_model_name_or_path = path)
+            vae = AutoencoderKL.from_pretrained(
+                pretrained_model_name_or_path=path)
             vae = None
         return True
     except:
         return False
 
-def diffusionGenerate_progress_callback(step :int, timestep :int, latents :torch.FloatTensor):
+
+def diffusionGenerate_progress_callback(step: int, timestep: int, latents: torch.FloatTensor):
     if generate_progress_callback is not None:
         generate_progress_callback(step, timestep, latents)
+
 
 async def diffusionGenerate_async(gc: GenerateContainer):
     global pipe
@@ -74,6 +81,6 @@ async def diffusionGenerate_async(gc: GenerateContainer):
         negative_prompt=gc.negative,
         num_images_per_prompt=gc.num,
         eta=gc.eta,
-        generator= None if gc.seed == -1 else torch.manual_seed(gc.seed),
+        generator=None if gc.seed == -1 else torch.manual_seed(gc.seed),
         callback=diffusionGenerate_progress_callback
     )).images
